moved database scripts to its own directory, started migrating to sequelize
