+++ package.json
@@ -25,7 +25,8 @@
     "dotenv": "^16.4.5",
     "ms": "^2.1.3",
     "nodemon": "^3.1.0",
-    "pg": "^8.12.0"
+    "pg": "^8.12.0",
+    "sequelize": "^6.37.3"
   },
   "devDependencies": {
     "@eslint/js": "^9.3.0",
@@ -34,6 +35,8 @@
     "globals": "^15.3.0"
   },
   "nodemonConfig": {
-    "ignore": ["**/*.json"]
+    "ignore": [
+      "**/*.json"
+    ]
   }
 }
