+++ src/database/index.js
@@ -1,0 +1,6 @@
+
+module.exports = {
+    ...require('./models'),
+    ...require('./queries'),
+    ...require('./errors')
+};
